change structure and names of some files
walk_buow_labels didn't make much sense, and i also wanted to
separate the functions for filtering the labels, as those are the
functions that will be varied based on the label file, whereas
ideally the other functions will be able to be the same across
data and label files. It will make it easier in the future for us
to add other parsers for different formats of label files
